// src/model/TrashModel.ts
export default interface TrashModel {
  DateRemoved: string;
  HostelId: string;
  StudentLevel: string;
  ReferenceNumber: string;
  Gender: string;
  DateBooked: string;
  Programme: string;
  Picture: string;
  AcademicYear: string;
  Email: string;
  IndexNumber: string;
  PhoneNumber: string;
  RoomNumber: string;
  StudentName: string;
}